major data arrangement changes in order to improve the button implementation. Button Implemented.
