+++ src/Components/Body.jsx
@@ -1,10 +1,16 @@
 import "../CSS/screenL-Body.Main.css";
 import HorizontalCarrier from "./HorizontalCarrier";
 
-const Body = () => {
+const Body = (props) => {
   return (
     <div className="mainBody-container">
-      <HorizontalCarrier />
+      <HorizontalCarrier
+        pfp={props.pfp}
+        name={props.name}
+        title1={props.title1}
+        title2={props.title2}
+        scrollMeTo={(where) => props.clickHandle(where)}
+      />
     </div>
   );
 };
