+++ src/Components/HorizontalCarrier.jsx
@@ -1,26 +1,23 @@
 import "../CSS/HCarrier.css";
 import IDCard from "./IDCard";
-import IDimage from "../images/PFP3.png";
 import HomeAnimation from "./homeAnimation";
-const personalData = {
-  pfp: IDimage,
-  name: "Armando Arias",
-  title1: "Front-End Developer",
-  title2: "UX / UI Designer",
-};
+import ScrollTo from "./ScrollTo.jsx";
 
-const HorizontalCarrier = () => {
+const HorizontalCarrier = (props) => {
   return (
     <div className="elements-carrier">
       <div id="home" className="spaceBlock">
         <IDCard
-          pfp={personalData.pfp}
-          name={personalData.name}
-          title1={personalData.title1}
-          title2={personalData.title2}
+          pfp={props.pfp}
+          name={props.name}
+          title1={props.title1}
+          title2={props.title2}
         />
         <div id="homeAnimation">
           <HomeAnimation />
+        </div>
+        <div className="aboutMeButton">
+          <ScrollTo name="About Me" whereTo={props.scrollMeTo("aboutMe")} />
         </div>
       </div>
       <div id="aboutMe" className="spaceBlock"></div>
