+++ src/Components/ScrollTo.jsx
@@ -1,0 +1,8 @@
+const ScrollTo = (props) => {
+  return (
+    <div>
+      <button onClick={props.whereTo}>{props.name}</button>
+    </div>
+  );
+};
+export default ScrollTo;
